requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
whitenoise==6.5.0
djangorestframework==3.14.0